// src/app/core/models/product.model.ts
export interface ProductModel {
        id: number;
        productId: string;
        productName: string;
        description: string;
        price: number;
        launchDate: string;
        photoUrl: string;
        isAvailable: boolean;
        isDeleted: boolean;
        catogoryName: string;
        catogory: string;
        productCart: string;
}